app houdt eerste en laatste machine per product bij, voor bijsturing productie; buffer heeft ook zijn next machines, zodat deze weet wanneer hij een laatste buffer is; application heeft een aantal functies erbij voor switchen van product (tijdelijk nu

+++ src/libs/models/MachineConfiguration.h
@@ -88,7 +88,7 @@ class MachineConfiguration {
    * Getter for initializationDuration in milliseconds
    * @return initializationDurationInSeconds * 1000
    */
-  uint16_t getInitializationDurationInMilliseconds() const;
+  uint32_t getInitializationDurationInMilliseconds() const;
 
   /**
    * Getter for inputBufferSize
